requirements.txt
asgiref==3.4.1
Django==3.2.9
django-admin-generator==2.2.1
django-crispy-forms==1.13.0
django-phonenumber-field==6.0.0
django-utils2==2.12.1
phonenumbers==8.12.37
pkg-resources==0.0.0
python-utils==2.5.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.0
